requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary    # PostgreSQL 드라이버
pydantic
python-dotenv      # .env 파일을 읽기 위한 라이브러리
